+++ CHANGELOG.md
@@ -5,6 +5,13 @@ More **detailed changelog** for each respective version may be viewed by pressin
 
 ## Version 1.x ##
 
+### [1.1.4](https://github.com/universum-studios/android_transitions/releases/tag/v1.1.4) ###
+> 27.02.2018
+
+- `Reveal`, `Scale`, `Translate` transitions now return **null** `Animator` from `createAnimator(...)`
+  factory methods and also from `onAppear(...)` and `onDisappear(...)` methods when a **target view**
+  to be animated is **not detached to window**.
+
 ### [1.1.3](https://github.com/universum-studios/android_transitions/releases/tag/v1.1.3) ###
 > 13.12.2017